🚚 QueryClient 로직 분리
